requirements.txt
numpy
pyyaml
opencv-python